django/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==24.4.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
isort==5.13.2
kombu==5.3.6
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
